Allow Trigger tag with alarm attribute from variable
