+++ core/src/main/java/nl/nn/adapterframework/monitoring/Clearing.java
@@ -3,7 +3,7 @@ package nl.nn.adapterframework.monitoring;
 /*
  * This class was created for the same reason that class Alarm was created.
  */
-public class Clearing extends Trigger {
+public class Clearing extends TriggerBase {
 	public Clearing() {
 		setAlarm(false);
 	}
