+++ core/src/main/java/nl/nn/adapterframework/monitoring/Monitor.java
@@ -64,7 +64,7 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 
 	private MonitorManager manager = null;
 
-	private List<Trigger> triggers = new ArrayList<>();
+	private List<TriggerBase> triggers = new ArrayList<>();
 	private Set<String> destinations = new HashSet<>(); 
 	private @Setter ApplicationContext applicationContext;
 
@@ -76,8 +76,8 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 		}
 
 		if (log.isDebugEnabled()) log.debug("monitor ["+getName()+"] configuring triggers");
-		for (Iterator<Trigger> it=triggers.iterator(); it.hasNext();) {
-			Trigger trigger = it.next();
+		for (Iterator<TriggerBase> it=triggers.iterator(); it.hasNext();) {
+			TriggerBase trigger = it.next();
 			if(!trigger.isConfigured()) {
 				trigger.configure();
 				((ConfigurableApplicationContext)applicationContext).addApplicationListener(trigger);
@@ -133,8 +133,8 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 	}
 
 	protected void clearEvents(boolean alarm) {
-		for (Iterator<Trigger> it=triggers.iterator(); it.hasNext();) {
-			Trigger trigger=(Trigger)it.next();
+		for (Iterator<TriggerBase> it=triggers.iterator(); it.hasNext();) {
+			TriggerBase trigger=(TriggerBase)it.next();
 			if (trigger.isAlarm()!=alarm) {
 				trigger.clearEvents();
 			}
@@ -162,8 +162,8 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 		monitor.addAttribute("name",getName());
 		monitor.addAttribute("type",getType());
 		monitor.addAttribute("destinations",getDestinationsAsString());
-		for (Iterator<Trigger> it=triggers.iterator();it.hasNext();) {
-			Trigger trigger=(Trigger)it.next();
+		for (Iterator<TriggerBase> it=triggers.iterator();it.hasNext();) {
+			TriggerBase trigger=(TriggerBase)it.next();
 			trigger.toXml(monitor);
 		}
 		return monitor;
@@ -214,12 +214,12 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 		}
 	}
 
-	public void registerTrigger(Trigger trigger) {
+	public void registerTrigger(TriggerBase trigger) {
 		trigger.setMonitor(this);
 		triggers.add(trigger);
 	}
 
-	public void removeTrigger(Trigger trigger) {
+	public void removeTrigger(TriggerBase trigger) {
 		int index = triggers.indexOf(trigger);
 		if(index > -1) {
 			AutowireCapableBeanFactory factory = applicationContext.getAutowireCapableBeanFactory();
@@ -246,10 +246,10 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 		return manager;
 	}
 
-	public List<Trigger> getTriggers() {
+	public List<TriggerBase> getTriggers() {
 		return triggers;
 	}
-	public Trigger getTrigger(int index) {
+	public TriggerBase getTrigger(int index) {
 		return triggers.get(index);
 	}
 
@@ -339,7 +339,7 @@ public class Monitor implements ApplicationContextAware, DisposableBean {
 		log.info("removing monitor ["+this+"]");
 
 		AutowireCapableBeanFactory factory = applicationContext.getAutowireCapableBeanFactory();
-		for (Trigger trigger : triggers) {
+		for (TriggerBase trigger : triggers) {
 			factory.destroyBean(trigger);
 		}
 	}
